Moved commands to the right place
Took 52 seconds
