+++ src/main/java/me/glaremasters/guilds/commands/CommandTransfer.java
@@ -22,65 +22,62 @@
  * SOFTWARE.
  */
 
-package me.glaremasters.guilds;
+package me.glaremasters.guilds.commands;
 
-import ch.jalu.configme.SettingsManager;
 import co.aikar.commands.ACFUtil;
 import co.aikar.commands.BaseCommand;
 import co.aikar.commands.annotation.CommandAlias;
+import co.aikar.commands.annotation.CommandCompletion;
 import co.aikar.commands.annotation.CommandPermission;
 import co.aikar.commands.annotation.Description;
+import co.aikar.commands.annotation.Single;
 import co.aikar.commands.annotation.Subcommand;
 import co.aikar.commands.annotation.Syntax;
+import co.aikar.commands.annotation.Values;
 import lombok.AllArgsConstructor;
-import me.glaremasters.guilds.configuration.sections.GuildSettings;
+import me.glaremasters.guilds.Messages;
 import me.glaremasters.guilds.exceptions.ExpectationNotMet;
 import me.glaremasters.guilds.exceptions.InvalidPermissionException;
 import me.glaremasters.guilds.guild.Guild;
-import me.glaremasters.guilds.guild.GuildHandler;
+import me.glaremasters.guilds.guild.GuildMember;
 import me.glaremasters.guilds.guild.GuildRole;
 import me.glaremasters.guilds.utils.Constants;
-import me.glaremasters.guilds.utils.StringUtils;
+import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
 
 /**
  * Created by Glare
  * Date: 4/5/2019
- * Time: 11:46 PM
+ * Time: 10:54 PM
  */
 @AllArgsConstructor @CommandAlias(Constants.ROOT_ALIAS)
-public class CommandRename extends BaseCommand {
-
-    private GuildHandler guildHandler;
-    private SettingsManager settingsManager;
+public class CommandTransfer extends BaseCommand {
 
     /**
-     * Rename a guild
-     * @param player the player renaming this guild
-     * @param guild the guild being renamed
+     * Transfer a guild to a new user
+     * @param player the player transferring this guild
+     * @param guild the guild being transferred
      * @param role the role of the player
-     * @param name new name of guild
+     * @param target the new guild master
      */
-    @Subcommand("rename")
-    @Description("{@@descriptions.rename}")
-    @CommandPermission(Constants.BASE_PERM + "rename")
-    @Syntax("<name>")
-    public void execute(Player player, Guild guild, GuildRole role, String name) {
-        if (!role.isChangeName())
+    @Subcommand("transfer")
+    @Description("{@@descriptions.transfer}")
+    @CommandPermission(Constants.BASE_PERM + "transfer")
+    @CommandCompletion("@members")
+    @Syntax("<player>")
+    public void execute(Player player, Guild guild, GuildRole role, @Values("@members") @Single String target) {
+        if (!role.isTransferGuild())
             ACFUtil.sneaky(new InvalidPermissionException());
 
-        if (!guildHandler.nameCheck(name, settingsManager))
-            ACFUtil.sneaky(new ExpectationNotMet(Messages.CREATE__REQUIREMENTS));
+        Player transfer = Bukkit.getPlayer(target);
 
-        if (settingsManager.getProperty(GuildSettings.BLACKLIST_TOGGLE)) {
-            if (guildHandler.blacklistCheck(name, settingsManager))
-                ACFUtil.sneaky(new ExpectationNotMet(Messages.ERROR__BLACKLIST));
-        }
+        if (transfer == null)
+            ACFUtil.sneaky(new ExpectationNotMet(Messages.ERROR__PLAYER_NOT_FOUND));
 
-        guild.setName(StringUtils.color(name));
+        guild.transferGuild(player, transfer);
 
-        getCurrentCommandIssuer().sendInfo(Messages.RENAME__SUCCESSFUL,
-                "{name}", name);
+        getCurrentCommandIssuer().sendInfo(Messages.TRANSFER__SUCCESS);
+        getCurrentCommandManager().getCommandIssuer(transfer).sendInfo(Messages.TRANSFER__NEWMASTER);
     }
 
 }
